services/underwriting-service/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
pydantic==2.6.4
psycopg2-binary==2.9.9
gunicorn==21.2.0 
PyYAML==6.0.1
requests==2.32.2